Correcion valor de cantidad en las tablas informeFinal y registros, avance en el modulo de registros

+++ talcos-express/models/InformeFinal.js
@@ -35,7 +35,7 @@ const InformeFinal = sequelize.define(
       allowNull: false,
     },
     cantidad_informe_final: {
-      type: DataTypes.BIGINT,
+      type: DataTypes.DECIMAL(5, 2),
       allowNull: false,
     },
     horometro_informe_final: {
